+++ packages/cli/src/commands/codegen.ts
@@ -10,22 +10,16 @@ import {
   parseManifestFileOption,
   parseClientConfigOption,
   getProjectFromManifest,
-  isPluginManifestLanguage,
-  generateWrapFile,
   defaultProjectManifestFiles,
   defaultPolywrapManifest,
   parseLogFileOption,
   parseWrapperEnvsOption,
 } from "../lib";
 import { ScriptCodegenerator } from "../lib/codegen/ScriptCodeGenerator";
+import { defaultCodegenDir } from "../lib/defaults/defaultCodegenDir";
 
 import { Env, PolywrapClient } from "@polywrap/client-js";
-import path from "path";
-import fs from "fs";
 import { Uri } from "@polywrap/core-js";
-
-const defaultCodegenDir = "./src/wrap";
-const defaultPublishDir = "./build";
 
 const pathStr = intlMsg.commands_codegen_options_o_path();
 const defaultManifestStr = defaultPolywrapManifest.join(" | ");
@@ -33,7 +27,6 @@ const defaultManifestStr = defaultPolywrapManifest.join(" | ");
 export interface CodegenCommandOptions extends BaseCommandOptions {
   manifestFile: string;
   codegenDir: string;
-  publishDir: string;
   script: string | false;
   clientConfig: string | false;
   wrapperEnvs: string | false;
@@ -58,12 +51,6 @@ export const codegen: Command = {
         })}`
       )
       .option(
-        `-p, --publish-dir <${pathStr}>`,
-        `${intlMsg.commands_codegen_options_publish({
-          default: defaultPublishDir,
-        })}`
-      )
-      .option(
         `-s, --script <${pathStr}>`,
         `${intlMsg.commands_codegen_options_s()}`
       )
@@ -83,16 +70,14 @@ export const codegen: Command = {
       )
       .action(async (options: Partial<CodegenCommandOptions>) => {
         await run({
-          ...options,
-          clientConfig: options.clientConfig || false,
-          wrapperEnvs: options.wrapperEnvs || false,
-          codegenDir: parseDirOption(options.codegenDir, defaultCodegenDir),
-          script: parseCodegenScriptOption(options.script),
           manifestFile: parseManifestFileOption(
             options.manifestFile,
             defaultProjectManifestFiles
           ),
-          publishDir: parseDirOption(options.publishDir, defaultPublishDir),
+          codegenDir: parseDirOption(options.codegenDir, defaultCodegenDir),
+          script: parseCodegenScriptOption(options.script),
+          clientConfig: options.clientConfig || false,
+          wrapperEnvs: options.wrapperEnvs || false,
           verbose: options.verbose || false,
           quiet: options.quiet || false,
           logFile: parseLogFileOption(options.logFile),
@@ -108,7 +93,6 @@ async function run(options: Required<CodegenCommandOptions>) {
     wrapperEnvs,
     codegenDir,
     script,
-    publishDir,
     verbose,
     quiet,
     logFile,
@@ -130,10 +114,13 @@ async function run(options: Required<CodegenCommandOptions>) {
   const project = await getProjectFromManifest(manifestFile, logger);
 
   if (!project) {
-    return;
+    logger.error(
+      `${intlMsg.commands_codegen_project_load_error({
+        manifestPath: manifestFile,
+      })}`
+    );
+    process.exit(1);
   }
-
-  const projectType = await project.getManifestLanguage();
 
   const schemaComposer = new SchemaComposer({
     project,
@@ -156,24 +143,6 @@ async function run(options: Required<CodegenCommandOptions>) {
       });
 
   const result = await codeGenerator.generate();
-
-  // HACK: Codegen outputs wrap.info into a build directory for plugins, needs to be moved into a build command?
-  if (isPluginManifestLanguage(projectType)) {
-    // Output the built manifest
-    const manifestPath = path.join(publishDir, "wrap.info");
-
-    if (!fs.existsSync(publishDir)) {
-      fs.mkdirSync(publishDir);
-    }
-
-    await generateWrapFile(
-      await schemaComposer.getComposedAbis(),
-      await project.getName(),
-      "plugin",
-      manifestPath,
-      logger
-    );
-  }
 
   if (result) {
     logger.info(`🔥 ${intlMsg.commands_codegen_success()} 🔥`);