+++ packages/js/plugins/uri-resolvers/file-system-resolver/src/__tests__/e2e.spec.ts
@@ -14,7 +14,7 @@ describe("Filesystem plugin", () => {
   let client: PolywrapClient;
 
   beforeAll(async () => {
-    await buildWrapper(simpleWrapperPath);
+    await buildWrapper(simpleWrapperPath, undefined, true);
 
     client = getClient();
   });