+++ packages/js/plugins/uri-resolvers/ipfs-resolver/src/__tests__/e2e.spec.ts
@@ -23,6 +23,7 @@ describe("IPFS Plugin", () => {
       ipfsProvider: providers.ipfs,
       ethereumProvider: providers.ethereum,
       ensName: "cool.wrapper.eth",
+      codegen: true
     });
 
     wrapperIpfsCid = ipfsCid;